requirements.txt
# FastAPI和Web服务
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# RAG核心组件
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.18

# 文档处理
pypdf2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
unstructured==0.11.6

# 数据处理和工具
numpy==1.24.3
pandas==2.1.4
requests==2.31.0
aiofiles==23.2.1
pydantic==2.5.2
pydantic-settings==2.1.0

# 开发和测试
pytest==7.4.3
black==23.11.0
flake8==6.1.0 